+++ package.json
@@ -4,12 +4,20 @@
   "scripts": {
     "test": "jest"
   },
-  "dependencies": {
-    "typescript": "^5.7.2"
-  },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.2.5"
-  }
+    "jest": "^29.7.0"
+  },
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/chan3199/javascript-calculator.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/chan3199/javascript-calculator/issues"
+  },
+  "homepage": "https://github.com/chan3199/javascript-calculator#readme",
+  "description": ""
 }
