+++ src/test/operator.test.js
@@ -1,0 +1,5 @@
+import { operationalCalculate } from "../js/operator";
+
+test("1 * 1 = 1", () => {
+  expect(operationalCalculate(1, 1, "+")).toBe(2);
+});
